code review tasks(#1)
