+++ src/test/java/com/axibase/tsd/api/method/sql/response/SqlApiResponseHeadersTests.java
@@ -1,6 +1,6 @@
 package com.axibase.tsd.api.method.sql.response;
 
-import com.axibase.tsd.api.method.sql.SqlTest;
+import com.axibase.tsd.api.method.sql.SqlMethod;
 import com.axibase.tsd.api.model.series.Sample;
 import com.axibase.tsd.api.model.series.Series;
 import org.junit.Assert;
@@ -18,7 +18,7 @@ import java.util.Set;
 /**
  * @author Igor Shmagrinslkiy
  */
-public class SqlApiResponseHeadersTests extends SqlTest {
+public class SqlApiResponseHeadersTests extends SqlMethod {
     private static final String TEST_PREFIX = "sql-response-headers";
     private static final String ACCESS_CONTROL_ALLOW_METHODS = "Access-Control-Allow-Methods";
     private static final String CONTENT_TYPE = "Content-type";
